+++ src/components/dialog/ImportProjectDialog.vue
@@ -1,0 +1,31 @@
+<template>
+  <default-dialog dialog-key="ImportProject" data-cy="import-project-dialog">
+    <template v-slot:title>
+      <q-icon color="primary" name="fa-brands fa-git-alt" />
+      {{ $t('page.home.importProject') }}
+    </template>
+    <template v-slot:default>
+      <import-project-form
+        @project:import="importProject"
+      />
+    </template>
+  </default-dialog>
+</template>
+
+<script setup>
+import DialogEvent from 'src/composables/events/DialogEvent';
+import DefaultDialog from 'src/components/dialog/DefaultDialog';
+import ImportProjectForm from 'src/components/form/ImportProjectForm';
+import { useRouter } from 'vue-router';
+
+const router = useRouter();
+
+/**
+ * Close importProjectDialog and redirect to project model page.
+ * @param {String} projectId - Id of project.
+ */
+function importProject(projectId) {
+  DialogEvent.next({ type: 'close', key: 'ImportProject' });
+  router.push(`/modelizer/${projectId}/model`);
+}
+</script>