+++ src/components/grid/ProjectGrid.vue
@@ -1,31 +1,52 @@
 <template>
   <div class="project-grid">
-    <q-table
-      v-if="Object.keys(projects).length !== 0"
-      :title="$t('projects.list')"
-      grid
-      :rows="Object.values(projects)"
-      row-key="id"
-      hide-header
-      hide-bottom
-      virtual-scroll
-      :pagination="{ rowsPerPage: 0 }"
-    >
-      <template v-slot:item="props">
-        <ProjectCard
-          class="q-pa-xs col-xs-12 col-sm-4 col-md-3"
-          :project="props.row"
+    <div>
+      <div class="row items-center">
+        <h4>{{ $t('page.home.recentProjects') }}</h4>
+        <q-btn
+          outline
+          no-caps
+          class="q-ml-xl"
+          color="primary"
+          icon="fa-solid fa-plus"
+          data-cy="new-project"
+          :label="$t('actions.home.newProject')"
+          @click="createProject"
         />
-      </template>
-    </q-table>
-    <div v-else caption class="text-center text-h6 text-grey empty-grid">
-      {{ $t('projects.empty') }}
+        <q-btn
+          outline
+          no-caps
+          class="q-ml-sm"
+          color="primary"
+          icon="fa-solid fa-cloud-arrow-down"
+          data-cy="import-project"
+          :label="$t('actions.home.importProject')"
+          @click="DialogEvent.next({ type: 'open', key: 'ImportProject' })"
+        />
+      </div>
+      <div class="row">
+        <ProjectCard
+          v-for="project in projects"
+          :key="project.id"
+          :project="project"
+          class="q-ma-sm"
+        />
+      </div>
+      <div
+        v-if="Object.keys(projects).length === 0"
+        class="row text-center text-h6 text-grey empty-grid"
+      >
+        {{ $t('projects.empty') }}
+      </div>
     </div>
   </div>
 </template>
 
 <script setup>
 import ProjectCard from 'src/components/card/ProjectCard.vue';
+import { useRouter } from 'vue-router';
+import { createProjectTemplate, initProject } from 'src/composables/Project';
+import DialogEvent from 'src/composables/events/DialogEvent';
 
 defineProps({
   projects: {
@@ -33,6 +54,16 @@ defineProps({
     required: true,
   },
 });
+const router = useRouter();
+
+/**
+ * Create project on fs, init project on git and redirect to project page.
+ */
+async function createProject() {
+  const project = createProjectTemplate();
+  await initProject(project);
+  router.push(`/modelizer/${project.id}/model`);
+}
 </script>
 
 <style scoped>