+++ tests/unit/components/card/ProjectCard.spec.js
@@ -25,4 +25,28 @@ describe('Test component: ProjectCard', () => {
       });
     });
   });
+
+  describe('Test functions', () => {
+    describe('Test function: getProjectImage', () => {
+      it('should return corresponding image name', async () => {
+        await wrapper.setProps({ project: { id: 'A' } });
+        expect(wrapper.vm.getProjectImage()).toEqual('images/project0.png');
+
+        await wrapper.setProps({ project: { id: 'B' } });
+        expect(wrapper.vm.getProjectImage()).toEqual('images/project1.png');
+
+        await wrapper.setProps({ project: { id: 'C' } });
+        expect(wrapper.vm.getProjectImage()).toEqual('images/project2.png');
+
+        await wrapper.setProps({ project: { id: 'D' } });
+        expect(wrapper.vm.getProjectImage()).toEqual('images/project3.png');
+
+        await wrapper.setProps({ project: { id: 'E' } });
+        expect(wrapper.vm.getProjectImage()).toEqual('images/project4.png');
+
+        await wrapper.setProps({ project: { id: 'F' } });
+        expect(wrapper.vm.getProjectImage()).toEqual('images/project0.png');
+      });
+    });
+  });
 });