+++ tests/unit/components/dialog/GitCommitDialog.spec.js
@@ -39,7 +39,7 @@ describe('Test component: GitCommitDialog', () => {
       },
       global: {
         plugins: [
-          createI18n(i18nConfiguration),
+          createI18n({ locale: 'en-US', messages: i18nConfiguration }),
         ],
       },
     });