+++ tests/unit/components/dialog/GitLogDialog.spec.js
@@ -49,7 +49,7 @@ describe('Test component: GitLogDialog', () => {
       },
       global: {
         plugins: [
-          createI18n(i18nConfiguration),
+          createI18n({ locale: 'en-US', messages: i18nConfiguration }),
         ],
       },
     });