+++ tests/unit/components/dialog/GitPushDialog.spec.js
@@ -31,7 +31,7 @@ describe('Test component: GitPushDialog', () => {
       },
       global: {
         plugins: [
-          createI18n(i18nConfiguration),
+          createI18n({ locale: 'en-US', messages: i18nConfiguration }),
         ],
       },
     });