+++ tests/unit/components/menu/GitBranchActionMenu.spec.js
@@ -34,7 +34,7 @@ describe('Test component: GitBranchActionMenu', () => {
       },
       global: {
         plugins: [
-          createI18n(i18nConfiguration),
+          createI18n({ locale: 'en-US', messages: i18nConfiguration }),
         ],
       },
     });